bugfix overflow

+++ 2020/A-Cardiology/main.cpp
@@ -1,22 +1,22 @@
 #include "bits/stdc++.h"
 using namespace std;
 
-int r,c;
+long r,c;
 int main(){
 	cin>>r>>c;
 	struct ansStru{
-		pair<int,int> rc;
+		pair<long,long> rc;
 		int iterationTimes;
 	};
 	vector<ansStru> stablePos(c);
-	for (int p = 0; p < c; ++p) {
+	for (long p = 0; p < c; ++p) {
 		//下标从0开始
-		pair<int,int> start={0,0},end={r-1,c-1};
+		pair<long,long> start={0,0},end={r-1,c-1};
 		int iterationTime=0;
-		pair<int,int> pthStart={r*p/c,r*p%c},
+		pair<long,long> pthStart={r*p/c,r*p%c},
 			pthEnd={(r*p+c-1)/c,(r*p+c-1)%c};
 		while(1){
-			auto rcAdd=[&](const pair<int,int>& rc,int v)->pair<int,int>{
+			auto rcAdd=[&](const pair<long,long>& rc,int v)->pair<long,long>{
 				auto ret=rc;
 				ret.first+=v/c;
 				v%=c;
@@ -34,14 +34,14 @@ int main(){
 				break;
 			++iterationTime;
 		}
-		stablePos[p].rc=(start == end ? start : pair<int, int>{-1, -1});
+		stablePos[p].rc=(start == end ? start : pair<long,long>{-1, -1});
 		stablePos[p].iterationTimes=iterationTime;
 	}
 	int ans,
 		minDist=999999999;
-	auto calDist=[&](const pair<int,int> rc)->int{
+	auto calDist=[&](const pair<long,long> rc)->long{
 		const int mr=r/2, mc=c/2;
-		int minr=abs(rc.first-mr),
+		long minr=abs(rc.first-mr),
 			minc=abs(rc.second-mc);
 		if(r%2)
 			minr=min(minr,abs(rc.first-mr+1));
@@ -52,7 +52,7 @@ int main(){
 	for (int i = 0; i < stablePos.size(); ++i) {
 		auto pi=stablePos[i].rc;
 		auto dist=calDist(pi);
-		if(pi!=pair<int,int>{-1,-1} && dist<minDist)
+		if(pi!=pair<long,long>{-1,-1} && dist<minDist)
 			minDist=dist,
 			ans=i;
 	}
